# public/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users",
                "operationId": "ddff855f47bf5ad43261171d0c6837dc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total_pages": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Out of range",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Out of range"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Find user by ID",
                "operationId": "e77734bee81bcadc9ef57e832a4cb2e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/reviews": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user's reviews",
                "operationId": "f915e61774b0fcb40d295a61740f5adc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user's reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Movie": {
                "title": "Movie model",
                "description": "Movie model",
                "properties": {
                    "id": {
                        "description": "ID of the movie",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "title": {
                        "description": "Title of the movie",
                        "type": "string",
                        "example": "The Godfather"
                    },
                    "running_time": {
                        "description": "Running time of the movie",
                        "type": "integer",
                        "example": 175
                    },
                    "cover_photo": {
                        "description": "Cover photo of the movie",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    },
                    "overview": {
                        "description": "Overview of the movie",
                        "type": "string",
                        "example": "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son."
                    },
                    "release_date": {
                        "description": "Release date of the movie",
                        "type": "string",
                        "example": "1972-03-14T00:00:00+00:00"
                    }
                },
                "type": "object"
            },
            "Review": {
                "title": "Review model",
                "description": "Review model",
                "properties": {
                    "id": {
                        "description": "ID of the review",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "movie": {
                        "$ref": "#/components/schemas/Movie"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "score": {
                        "description": "Score of the review",
                        "type": "integer",
                        "example": 5
                    },
                    "comment": {
                        "description": "Comment of the review",
                        "type": "string",
                        "example": "This movie was amazing!"
                    },
                    "date_reviewed": {
                        "description": "Date the review was written",
                        "type": "string",
                        "example": "2022-01-01T00:00:00+00:00"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "username": {
                        "description": "Username of the user",
                        "type": "string",
                        "example": "john_doe"
                    },
                    "restricted": {
                        "description": "Whether the user is restricted",
                        "type": "boolean",
                        "example": false
                    },
                    "date_joined": {
                        "description": "Date the user joined",
                        "type": "string",
                        "example": "2022-01-01T00:00:00+00:00"
                    },
                    "profile_image": {
                        "description": "Profile image of the user",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    }
}